backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6
sqlmodel>=0.0.14
numpy>=1.26.3
scipy>=1.11.4
python-dotenv>=1.0.0
deepgram-sdk>=5.3.1
pyaudio>=0.2.14
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
pyttsx3>=2.90
openai>=1.0.0

